Changing the way to make empty arrays.
Instead of using empty parantheses on an intented line, the argument should now simply not be specified. This is more consistent with the existing syntax because every indented line should be an element in an array, even if it only contained empty parantheses. So the previous syntax should have resulted in an array with a single element, not an empty array. Since all arguments should be arrays, a nonexisting argument can be interpreted as an empty array. This also makes empty arrays an alternative to using null, which can be considered an undesirable element in a programming language.